refactor(#step2): 기타 리팩토링
