+++ src/main/java/kitchenpos/eatinorders/ui/OrderRestController.java
@@ -22,7 +22,7 @@ public class OrderRestController {
     public ResponseEntity<Order> create(@RequestBody final Order request) {
         final Order response = orderService.create(request);
         return ResponseEntity.created(URI.create("/api/orders/" + response.getId()))
-            .body(response);
+                .body(response);
     }
 
     @PutMapping("/{orderId}/accept")
