+++ src/main/java/kitchenpos/menus/dto/MenuCreateRequest.java
@@ -38,5 +38,5 @@ public final class MenuCreateRequest {
     public List<MenuProductElement> getMenuProducts() {
         return menuProducts;
     }
-    
+
 }
