+++ src/main/java/kitchenpos/products/tobe/domain/ProductDomainService.java
@@ -1,6 +1,6 @@
 package kitchenpos.products.tobe.domain;
 
-import kitchenpos.support.product.event.ProductPriceChangedEvent;
+import kitchenpos.products.event.ProductPriceChangedEvent;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
