+++ src/main/resources/db/migration/V1__Initialize_project_tables.sql
@@ -1,24 +1,24 @@
 create table menu
 (
-    id            binary(16)     not null,
+    id            binary(16) not null,
     displayed     bit            not null,
     name          varchar(255)   not null,
     price         decimal(19, 2) not null,
-    menu_group_id binary(16)     not null,
+    menu_group_id binary(16) not null,
     primary key (id)
 ) engine = InnoDB;
 
 create table menu_group
 (
-    id   binary(16)   not null,
+    id   binary(16) not null,
     name varchar(255) not null,
     primary key (id)
 ) engine = InnoDB;
 
 create table menu_product
 (
-    seq        bigint     not null auto_increment,
-    quantity   bigint     not null,
+    seq        bigint not null auto_increment,
+    quantity   bigint not null,
     product_id binary(16) not null,
     menu_id    binary(16) not null,
     primary key (seq)
@@ -26,8 +26,8 @@ create table menu_product
 
 create table order_line_item
 (
-    seq      bigint     not null auto_increment,
-    quantity bigint     not null,
+    seq      bigint not null auto_increment,
+    quantity bigint not null,
     menu_id  binary(16) not null,
     order_id binary(16) not null,
     primary key (seq)
@@ -35,7 +35,7 @@ create table order_line_item
 
 create table order_table
 (
-    id               binary(16)   not null,
+    id               binary(16) not null,
     occupied         bit          not null,
     name             varchar(255) not null,
     number_of_guests integer      not null,
@@ -44,9 +44,9 @@ create table order_table
 
 create table orders
 (
-    id               binary(16)   not null,
+    id               binary(16) not null,
     delivery_address varchar(255),
-    order_date_time  datetime(6)  not null,
+    order_date_time  datetime(6) not null,
     status           varchar(255) not null,
     type             varchar(255) not null,
     order_table_id   binary(16),
@@ -55,7 +55,7 @@ create table orders
 
 create table product
 (
-    id    binary(16)     not null,
+    id    binary(16) not null,
     name  varchar(255)   not null,
     price decimal(19, 2) not null,
     primary key (id)
