+++ src/test/java/kitchenpos/menus/tobe/domain/menu/FakeMenu.java
@@ -9,18 +9,18 @@ import java.util.UUID;
 
 public class FakeMenu extends Menu {
 
-        public static Menu createFake(String name, PurgomalumClient purgomalumClient, long price, MenuGroup menuGroup, boolean displayed, List<MenuProduct> menuProducts) {
-            MenuPrice menuPrice = new MenuPrice();
-            menuPrice.value = BigDecimal.valueOf(price);
+    public static Menu createFake(String name, PurgomalumClient purgomalumClient, long price, MenuGroup menuGroup, boolean displayed, List<MenuProduct> menuProducts) {
+        MenuPrice menuPrice = new MenuPrice();
+        menuPrice.value = BigDecimal.valueOf(price);
 
-            return new Menu(
-                    UUID.randomUUID(),
-                    MenuName.create(name, purgomalumClient),
-                    menuPrice,
-                    menuGroup,
-                    MenuDisplay.create(displayed),
-                    menuProducts
-            );
-        }
+        return new Menu(
+                UUID.randomUUID(),
+                MenuName.create(name, purgomalumClient),
+                menuPrice,
+                menuGroup,
+                MenuDisplay.create(displayed),
+                menuProducts
+        );
+    }
 
 }
