+++ src/test/java/kitchenpos/products/tobe/domain/ProductDomainServiceTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.products.tobe.domain;
 
-import kitchenpos.support.product.event.ProductPriceChangedEvent;
+import kitchenpos.products.event.ProductPriceChangedEvent;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
